SCSS Rewrite (pt 1)
It was messy and gross!
Now it's SCSS, cleaner, and looks more like what the styles in purgatory
are supposed to look like. How nice! I removed a lot of the useless
classes that were in the original style—no rules were declared for them
anyway.

index.html is fully good to go! (minus mobile compatibility)